% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{analysesOutputsTimes}
\alias{analysesOutputsTimes}
\title{Simulation timesteps for analyses}
\usage{
analysesOutputsTimes(period, interval)
}
\arguments{
\item{period}{numeric vector of length 2 corresponding to the start and end times
to use for analyses.}

\item{interval}{numeric indicating the interval between timesteps for analyses}
}
\value{
numeric vector of timesteps for which to run analyses
}
\description{
Simulation timesteps for analyses
}
