% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{cleanAreaName}
\alias{cleanAreaName}
\title{Extract study area name from run name}
\usage{
cleanAreaName(area)
}
\arguments{
\item{area}{Simulated area (i.e., run) name}
}
\description{
Extract study area name from run name
}
