% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studyAreas.R
\name{extractFMA}
\alias{extractFMA}
\title{Extract boundary polygon(s) for LandWeb forest management area(s)}
\usage{
extractFMA(fmas, name)
}
\arguments{
\item{fmas}{\code{sf} object with LandWeb FMA boundaries}

\item{name}{A character (regex) string to match.}
}
\value{
\code{sf} polygons object
}
\description{
Extract boundary polygon(s) for LandWeb forest management area(s)
}
