% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studyAreas.R
\name{extractFMU}
\alias{extractFMU}
\title{Extract boundary polygon(s) for Alberta forest management unit(s)}
\usage{
extractFMU(fmus, name)
}
\arguments{
\item{fmus}{\code{sf} object with Alberta FMU boundaries}

\item{name}{A character (regex) string to match.}
}
\value{
\code{sf} polygons object
}
\description{
Extract boundary polygon(s) for Alberta forest management unit(s)
}
