% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportingPolygons.R
\name{joinReportingPolygons}
\alias{joinReportingPolygons}
\title{Join reporting polygons and intersect their features}
\usage{
joinReportingPolygons(x, y)
}
\arguments{
\item{x, y}{a \code{sf}}
}
\value{
an \code{sf} polygons object
}
\description{
Join two reporting polygons, preserving their features;
concatenate the \verb{Name.*} fields into a single \code{Name} field.
E.g., if \code{x} and \code{y} each contain 2 features, the resulting object will contain 4
features (corresponding to \code{x1.y1}, \code{x1.y2}, \code{x2.y1}, and \code{x2.y2}).
}
