% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landmine-helpers.R
\name{landmine_optim_clusterExport}
\alias{landmine_optim_clusterExport}
\title{Export objects used for optimization and load packages on cluster}
\usage{
landmine_optim_clusterExport(cl = NULL, objs = NULL, pkgs = NULL)
}
\arguments{
\item{cl}{a cluster object or \code{NULL}}

\item{objs}{character vector of names of objects to export}

\item{pkgs}{character vector of packages to pre-load on the cluster nodes}
}
\value{
\code{NULL}. Invoked for its side effects.
}
\description{
Export objects used for optimization and load packages on cluster
}
