% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landmine-helpers.R
\name{landmine_optim_clusterWrap}
\alias{landmine_optim_clusterWrap}
\title{Wrapper function to setup cluster, export objects and load packages}
\usage{
landmine_optim_clusterWrap(cl = NULL, nodes, reps, objs, pkgs)
}
\arguments{
\item{cl}{a cluster object or \code{NULL}}

\item{nodes}{positive integer of length 1 specifying the number of threads
to use on the current machine (\code{localhost}), or a character vector of
hostnames on which to run worker copies.}

\item{reps}{integer. number of replicates to run.}

\item{objs}{character vector of names of objects to export}

\item{pkgs}{character vector of packages to pre-load on the cluster nodes}
}
\value{
named list of length 2 containing:
\code{cl}, a cluster object;
\code{out}, a list of burn maps (aka \code{burnMapList}).
}
\description{
Wrapper function to setup cluster, export objects and load packages
}
