% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landmine-calibration.R
\name{landmine_optim_islandTarget}
\alias{landmine_optim_islandTarget}
\title{Andison island-remnant target}
\usage{
landmine_optim_islandTarget(fireAreaHa)
}
\arguments{
\item{fireAreaHa}{Fire area, in hectares.}
}
\value{
Expected percent of fire area in remnant islands.
}
\description{
Percent of total fire area expected to be in remnant islands, by disturbance size class
(Andison 1996, Table 3.5, \strong{empirical} column from DeLong and Tanner 1995). The model column
of that table (3.3 / 6.2 / 10.5) is the previous model's output, not a target.
}
