% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landweb-area.R
\name{landweb_area}
\alias{landweb_area}
\title{The outline of the LandWeb area}
\usage{
landweb_area(lthfc)
}
\arguments{
\item{lthfc}{the cleaned LTHFC polygons: \code{\link[=polygonClean]{polygonClean()}} of \code{\link[=landweb_lthfc]{landweb_lthfc()}}.}
}
\value{
\code{sfc} polygon: the union of \code{lthfc}, made valid, without holes.
}
\description{
The outline of the LandWeb area
}
