% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LandWebUtils-package.R
\docType{package}
\name{LandWebUtils-package}
\alias{LandWebUtils}
\alias{LandWebUtils-package}
\title{LandWebUtils: Helper functions for the LandWeb project}
\description{
Additional utilities for LandWeb analyses.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://LandWebUtils.predictiveecology.org}
  \item \url{https://github.com/PredictiveEcology/LandWebUtils}
}

}
\author{
\strong{Maintainer}: Alex M Chubaty \email{achubaty@for-cast.ca} (\href{https://orcid.org/0000-0001-7146-8135}{ORCID})

Authors:
\itemize{
  \item Alex M Chubaty \email{achubaty@for-cast.ca} (\href{https://orcid.org/0000-0001-7146-8135}{ORCID})
  \item Eliot J B McIntire \email{eliot.mcintire@NRCan-RNCan.gc.ca} (\href{https://orcid.org/0000-0002-6914-8316}{ORCID})
}

Other contributors:
\itemize{
  \item His Majesty the Queen in Right of Canada, as represented by the Minister of Natural Resources Canada [copyright holder]
}

}
\keyword{internal}
