% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CASFRIv4.R
\name{loadCASFRI}
\alias{loadCASFRI}
\title{Load CASFRI data for LandWeb}
\usage{
loadCASFRI(
  CASFRIRas,
  attrFile,
  headerFile,
  sppEquiv,
  sppEquivCol,
  type = c("cover", "age")
)
}
\arguments{
\item{CASFRIRas}{TODO: description needed}

\item{attrFile}{TODO: description needed}

\item{headerFile}{TODO: description needed}

\item{sppEquiv}{table with species name equivalencies between the kNN and final naming formats.
See \code{data("sppEquivalencies_CA", "LandR")}.
For functions that have \code{mixedType} e.g., \code{vegTypeMapGenerator},
this only necessary if \code{mixedType == 2}.
If not provided and \code{mixedType == 2}, will attempt to use
\code{data("sppEquivalencies_CA", "LandR")}.}

\item{sppEquivCol}{the column name to use from \code{sppEquiv}.
For functions that have \code{mixedType} e.g., \code{vegTypeMapGenerator},
Only necessary if \code{mixedType == 2}.
If not provided and \code{mixedType == 2}, will attempt to use "Boreal".}

\item{type}{Character string. Either \code{"cover"} or \code{"age"}.}
}
\value{
TODO: description needed
}
\description{
Using CASFRI v4 data prepared for LandWeb.
}
