% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landmine-helpers.R
\name{meanTruncPareto}
\alias{meanTruncPareto}
\title{Calculate the mean of a truncated Pareto distribution}
\usage{
meanTruncPareto(k, lower, upper, alpha)
}
\arguments{
\item{k}{TODO: description needed}

\item{lower}{TODO: description needed}

\item{upper}{TODO: description needed}

\item{alpha}{TODO: description needed}
}
\value{
TODO: description needed
}
\description{
Calculate the mean of a truncated Pareto distribution
}
