% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotFMA}
\alias{plotFMA}
\alias{plotLandWeb}
\alias{plotGG}
\title{Plot boundary polygon(s) for forest management areas}
\usage{
plotFMA(x, provs, caribou = NULL, xsr = NULL, title = NULL, png = NULL)

plotLandWeb(x, provs, caribou = NULL, xsr = NULL, title = NULL, png = NULL)

plotGG(x, provs, caribou = NULL, png = NULL, title = NULL)
}
\arguments{
\item{x}{\code{sf} object corresponding to the FMA to be plotted}

\item{provs}{\code{sf} object corresponding to the provincial (or territorial) boundaries to plot}

\item{caribou}{Optional \code{sf} object corresponding to caribou boundaries}

\item{xsr}{Optional \code{sf} object corresponding to a buffered \code{studyArea}}

\item{title}{Character string to use for plot title}

\item{png}{Optional. If non-NULL, must be a valid file path to write a png}
}
\description{
Plot boundary polygon(s) for forest management areas
}
