% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportingPolygons.R
\name{prepFMAs}
\alias{prepFMAs}
\alias{prepFMUs}
\alias{prepANSRs}
\alias{prepEcoregions}
\alias{prepEcoprovinces}
\title{Prepare reporting polygons}
\usage{
prepFMAs(destinationPath, targetCRS = LandWebCRS)

prepFMUs(destinationPath, targetCRS = LandWebCRS)

prepANSRs(destinationPath, targetCRS = LandWebCRS)

prepEcoregions(destinationPath, targetCRS = LandWebCRS)

prepEcoprovinces(destinationPath, targetCRS = LandWebCRS)
}
\arguments{
\item{destinationPath}{character specifying path to destination directory}

\item{targetCRS}{character specifying the CRS to reproject polygons to}
}
\description{
\itemize{
\item LandWeb FMA boundaries (FMAs);
\item Alberta FMU boundaries (FMUs);
\item Alberta Natural Subregions (ANSRs);
}
}
