% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studyAreas.R
\name{prepStudyArea}
\alias{prepStudyArea}
\alias{prepRandomStudyArea}
\title{Extract boundary polygon(s) for LandWeb study areas}
\usage{
prepStudyArea(name, destinationPath, targetCRS = LandWebCRS)

prepRandomStudyArea(destinationPath, targetCRS = LandWebCRS, .seed = 867)
}
\arguments{
\item{name}{A character (regex) string to match.}

\item{destinationPath}{character specifying path to destination directory}

\item{targetCRS}{character specifying the CRS to reproject polygons to}
}
\value{
\code{sf} polygons object
}
\description{
Extract boundary polygon(s) for LandWeb study areas
}
