% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplots.R
\name{runBoxPlotsVegCover}
\alias{runBoxPlotsVegCover}
\title{Generate box and whisker plots for leading vegetation cover}
\usage{
runBoxPlotsVegCover(map, functionName, analysisGroups, dPath)
}
\arguments{
\item{map}{A \code{map} object.}

\item{functionName}{TODO: description needed}

\item{analysisGroups}{TODO: description needed}

\item{dPath}{Destination path for the resulting PNG files.}
}
\description{
Generate box and whisker plots for leading vegetation cover
}
