% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histograms.R
\name{runHistsVegCover}
\alias{runHistsVegCover}
\title{Generate histograms for leading vegetation cover}
\usage{
runHistsVegCover(map, functionName, analysisGroups, dPath)
}
\arguments{
\item{map}{A \code{map} object.}

\item{functionName}{Character specifying the name of the function to apply.}

\item{analysisGroups}{Character specifying the analysis groups to run.}

\item{dPath}{Destination path for the resulting PNG files.}
}
\description{
Generate histograms for leading vegetation cover
}
