% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{simFile}
\alias{simFile}
\title{Generate simulation file name}
\usage{
simFile(name, path, time)
}
\arguments{
\item{name}{Object name (e.g., \code{"mySimOut"})}

\item{path}{Directory location in where the file will be located (e.g., an \code{outputPath}).}

\item{time}{Optional simulation time to use as filename suffix.}
}
\description{
Assists with saving/retrieving LandWeb simulations.
}
