% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simfiles.R
\name{findSimFile}
\alias{findSimFile}
\alias{findOldSimFiles}
\title{Find LandWeb simulation output files}
\usage{
findSimFile(outputDir = NULL, rep = NULL)

findOldSimFiles(outputDir = NULL, rep = NULL, before = "2024-01-01")
}
\arguments{
\item{outputDir}{path to LandWeb output directory.}

\item{rep}{integer giving the replicate id, or character string in the form of \code{"rep01"}.}

\item{before}{date before which files are considered outdated (default "2024-01-01").}
}
\value{
path to the file
}
\description{
Find LandWeb simulation output files
}
\examples{
\dontrun{
  outDir <- file.path("~/GitHub/LandWeb/outputs/Tolko_AB_S_aspenDispersal_logROS")
  if (dir.exists(outdir)) {
    oldFiles <- findOldSimFiles(outDir)      ## search all reps
    oldFiles <- findOldSimFiles(outDir, 1:5) ## search specific reps
    # fs::file_delete(oldFiles) ## double check before deleting
  }
}

}
