% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speciesTable.R
\name{updateSpeciesTable}
\alias{updateSpeciesTable}
\title{Customize species trait table values for LandWeb}
\usage{
updateSpeciesTable(speciesTable, runName)
}
\arguments{
\item{speciesTable}{The species traits table}

\item{runName}{The LandWeb run name}
}
\description{
Customize species trait table values for LandWeb
}
