% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{findSimFile}
\alias{findSimFile}
\title{Find LandWeb simulation output file}
\usage{
findSimFile(outputDir, area, rep)
}
\arguments{
\item{outputDir}{path to LandWeb output directory.}

\item{area}{character string giving study area with run name.}

\item{rep}{integer giving the replicate id, or character string in the form of \code{"rep01"}.}
}
\value{
path to the file
}
\description{
Find LandWeb simulation output file
}
