% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landmine.R
\name{landmine_burn1}
\alias{landmine_burn1}
\alias{landmine_burn}
\title{Core Burn function for Andison's LandMine Fire Module}
\usage{
landmine_burn1(
  landscape,
  startCells,
  fireSizes = 5,
  nActiveCells1 = c(10, 36),
  spawnNewActive = c(0.46, 0.2, 0.26, 0.11),
  maxRetriesPerID = 10L,
  sizeCutoffs = c(8000, 20000),
  spreadProbRel = spreadProbRel,
  spreadProb = 0.77
)

landmine_burn(
  landscape,
  startCells,
  fireSizes = 5,
  nActiveCells1 = c(10, 36),
  spawnNewActive = c(0.46, 0.2, 0.26, 0.11),
  sizeCutoffs = c(8000, 20000),
  spreadProbRel = 0.23
)
}
\arguments{
\item{landscape}{A \code{RasterLayer}. This only provides the extent and
resolution for the fire spread algorithm.}

\item{startCells}{A numeric vector indicating the indices on the \code{landscape}
where the fires will start with 100\%\% certainty.}

\item{fireSizes}{A numeric vector indicating the final size of each of the fires.
Must be same length as \code{startCells}.}

\item{nActiveCells1}{A numeric vector of length 2. These are cutoffs above and
below each of which different values of \code{spawnNewActive}
are used. See details.}

\item{spawnNewActive}{A numeric vector of length 4. These are the probabilities
of creating spreading to 2 neighbours instead of the 1
default neighbour, each time step.
The 4 values are for 4 different fire size conditions. See details.}

\item{maxRetriesPerID}{Integer. Maximum number of retry attempts per firelet ID.}

\item{sizeCutoffs}{A numeric vector of length 2.
These are 2 size (in hectares) thresholds that affect which
\code{spawnNewActive} probabilities are used. See details.}

\item{spreadProbRel}{A raster layer of of relative probabilities, with non-flammable pixels \code{NA}.}

\item{spreadProb}{A raster layer of spread probabilities, with non-flammable pixels \code{NA}.}
}
\value{
A \code{data.table} with 4 columns
}
\description{
The main function for the Andison Fire Module. See details.
}
\details{
This algorithm is a modified contagious cellular automaton.
}
\note{
The original version (\code{landmine_burn()}) is deprecated and should not be used.
Use \code{landmine_burn1()} instead.
}
\section{Algorithm}{


\subsection{Core}{
Each fire starts at a single pixel, \code{startCells} and will spread,
i.e., visit and convert from a 0 to the fire id number.
It will iteratively spread until the number of cells visited is equal to \code{floor(fireSizes)}.
}

\subsection{Adjustments due to current fire size and number of active pixels}{
That can vary too, but it gets a bit complicated, so use that for now.
Spawning probability was originally set at 13\%, but created problems with very
large and very small fires, so over time has been adjusted to vary depending on:
a) number of active "firelets" (NF); and b) fire size (FS), such that:

\if{html}{\out{<div class="sourceCode">}}\preformatted{  - If 10 <= NF < 36 and FS < 20,000 ha then P = 20\%
  - If NF > 36 and FS < 8,000 ha, P = 11\%
  - If NF < 36 and FS > 20,000 ha, P = 26\%
  - If NF < 10 then P = 46\%
}\if{html}{\out{</div>}}

These rule create more heterogeneity in the pattern of burning.
}

\subsection{Fire jumping}{
If the fire has not reached its target size, it will try to pick new neighbours among
the 8 immediate neighbours up to 4 times.
If it still did not find enough neighbours, then it will jump or "spot" up to 4 pixels away.
It will then repeat the previous 2 stages again once (i.e., 4 neighbours, 1 jump, repeat),
then it will stop, unable to achieve the desired \code{fireSize}.
}
}

