% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landmine-helpers.R
\name{landmine_fitSN}
\alias{landmine_fitSN}
\alias{fitSN2}
\title{LandMine objective functions}
\usage{
landmine_fitSN(
  sna,
  ros,
  centreCell,
  fireSizes = 10^(2:5),
  desiredPerimeterArea = 0.004
)

fitSN2(
  par,
  ros,
  centreCell,
  fireSizes = 10^(2:5),
  desiredPerimeterArea = 0.003,
  spreadProb = 0.9
)
}
\arguments{
\item{sna}{TODO}

\item{ros}{TODO}

\item{centreCell}{TODO}

\item{fireSizes}{TODO}

\item{desiredPerimeterArea}{TODO}

\item{par}{parameter vector of length 5}

\item{spreadProb}{spread probability}
}
\value{
\code{data.table} (TODO)
}
\description{
\code{landmine_fitSN()} is used for the module.
}
\details{
\code{landmine_fitSN2()} is an alternative version tries the optimization using fewer parameters,
to test whether a simpler version gets better/different results.
Although this version was not used for the final module, we preserve it here for posterity.
}
