% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landmine-helpers.R
\name{landmine_optim_burnFun}
\alias{landmine_optim_burnFun}
\title{LandMine burn optimization function}
\usage{
landmine_optim_burnFun(
  ros,
  centreCell,
  fireSize,
  spawnNewActive,
  sizeCutoffs,
  spreadProb
)
}
\arguments{
\item{ros}{\code{RasterLayer} of LandMine Raster Of Spread values}

\item{centreCell}{TODO}

\item{fireSize}{TODO}

\item{spawnNewActive}{TODO}

\item{sizeCutoffs}{TODO}

\item{spreadProb}{TODO}
}
\value{
named list of length 2 containing:
\code{burnedMap}: \code{rasterLayer} of burned pixels;
\code{LM}: \code{data.frame} of patch statistics from \code{SDMTools::PatchStats()}.
}
\description{
LandMine burn optimization function
}
