% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landmine-helpers.R
\name{landmine_optim_clusterSetup}
\alias{landmine_optim_clusterSetup}
\title{Setup a cluster for LandMine optimization}
\usage{
landmine_optim_clusterSetup(nodes = NULL)
}
\arguments{
\item{nodes}{positive integer of length 1 specifying the number of threads
to use on the current machine (\code{localhost}), or a character vector of
hostnames on which to run worker copies.}
}
\value{
a cluster objct
}
\description{
Setup a cluster for LandMine optimization
}
