% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landmine-plot.R
\name{landmine_plots}
\alias{landmine_plots}
\alias{landmine_plot_areaBurnedOverTime}
\alias{landmine_plot_LTHFC}
\alias{landmine_plot_FRI}
\title{LandMine diagnostic plots}
\usage{
landmine_plot_areaBurnedOverTime(areaBurnedOverTime)

landmine_plot_LTHFC(lthfc, studyAreaName, ...)

landmine_plot_FRI(friSummary)
}
\arguments{
\item{areaBurnedOverTime}{Summary \code{data.frame} of area burned over time, containing
the following columns:
\code{time} (numeric) gives the simulation time (year);
\code{haBurned} (numeric) gives the burned area in hectares;
\code{FRI} (numeric) identifies the fire return interval polygon.}

\item{lthfc}{long-term historic fire cycle map (raster).}

\item{studyAreaName}{study area name (character).}

\item{...}{additional arguments passed to \code{\link[rasterVis:levelplot-methods]{rasterVis::levelplot()}}.}

\item{friSummary}{Summary \code{data.frame} of simulated fire return intervals (FRI) vs.
the long-term historic fire cycles (LTHFC), containing the following columns:
\code{simArea} (character) gives the study area name;
\code{LTHFC} (numeric) is the expected (i.e., historic) FRI;
\code{FRI} (numeric) is the simulated FRI.}
}
\value{
a \code{ggplot2} or \code{rasterVis} object; invoked for side effect of creating plots.
}
\description{
\itemize{
\item \code{landmine_plot_areaBurnedOverTime()} plots the area burned over time by LTHFC polygon;
\item \code{landmine_plot_LTHFC()} produces a \code{\link[rasterVis:levelplot-methods]{rasterVis::levelplot()}} a map of the LTHFC polygons;
\item \code{landmine_plot_FRI()} plots
}
}
