% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportingPolygons.R
\docType{data}
\name{LandWebCRS}
\alias{LandWebCRS}
\title{Target CRS (projection) to use with LandWeb}
\format{
An object of class \code{character} of length 1.
}
\usage{
LandWebCRS
}
\description{
As of September 2025, this corresponds to the SCANFI's \code{NAD_1983_Canada_Lambert} projection.
}
\note{
needs to be character, not \code{CRS} class, for downstream use with \code{data.table}
}
\keyword{datasets}
