% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studyAreas.R
\docType{data}
\name{LandWebStudyAreas}
\alias{LandWebStudyAreas}
\alias{studyAreaProv}
\alias{studyAreaIn}
\title{Study areas available for LandWeb simulations}
\format{
An object of class \code{data.frame} with 32 rows and 4 columns.
}
\usage{
LandWebStudyAreas

studyAreaProv(name)

studyAreaIn(name, prov)
}
\arguments{
\item{name}{character specifying the study area name. Must be one of \code{LandWebStudyAreas$Name}.}

\item{prov}{character specifying the province abbreviation (i.e., "AB", "BC", "MB", "NWT", or "SK").}
}
\value{
\code{LandWebStudyAreas} is a data.frame` with 4 columns:
\itemize{
\item \code{Name} specifies the shortname of the study area for use with simulation setup;
\item \code{Province} identifies the location of the study area;
\item \code{ID} corresponds to the polygon ID of the Updated FMA Boundaries polygons;
\item \code{Description} provides the full study area name and additional context (if any);
}

\code{studyAreaProv} returns a character indicating the province the study area is in.

\code{studyAreaIn} returns a logical indicating whether study area \code{name} is in \code{prov}.
}
\description{
Study areas available for LandWeb simulations
}
\keyword{datasets}
