% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/largePatches.R
\name{LargePatches}
\alias{LargePatches}
\title{Calculate proportion of large patches in NRV}
\usage{
LargePatches(
  tsf,
  vtm,
  poly,
  labelColumn,
  id,
  ageClassCutOffs,
  ageClasses,
  sppEquivCol,
  sppEquiv,
  crop2poly = FALSE
)
}
\arguments{
\item{tsf}{A single filename, relative or absolute, pointing to a Time Since Fire raster.
Can be any format that \code{raster} can use.}

\item{vtm}{A single filename, relative or absolute, pointing to a Vegetation Type Map raster.
Can be any format that \code{raster} can use.}

\item{poly}{A single \code{SpatialPolygonsDataFrame} object or a factor \code{RasterLayer}.
This layer MUST have a column labelled \code{shinyLabel}}

\item{labelColumn}{TODO: description needed}

\item{id}{TODO: description needed}

\item{ageClassCutOffs}{A numeric vector with the endpoints for the \code{ageClasses}.
Should be \code{length(ageClasses) + 1}. See \code{.ageClassCutOffs}.}

\item{ageClasses}{A character vector with labels for age classes to bin the \code{tsf} times,
e.g., \code{c("Young", "Immature", "Mature", "Old")}. See \code{.ageClasses}.}

\item{sppEquivCol}{Character giving the column name to use in \code{sppEquiv}.}

\item{sppEquiv}{Species equivalency table, e.g., derived from \code{LandR::sppEquivalencies_CA}.}

\item{crop2poly}{logical indicating whether to crop/mask \code{vtm} and \code{tsf}
rasters to \code{poly}. Default \code{FALSE} for backwards compatibility.}
}
\description{
TODO: needs description
}
