% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/largePatches.R
\name{LargePatches}
\alias{LargePatches}
\title{Calculate proportion of large patches in NRV}
\usage{
LargePatches(
  tsf,
  vtm,
  poly,
  labelColumn,
  id,
  ageClassCutOffs,
  ageClasses,
  sppEquivCol,
  sppEquiv,
  crop2poly = FALSE
)
}
\arguments{
\item{tsf}{TODO: description needed}

\item{vtm}{TODO: description needed}

\item{poly}{TODO: description needed}

\item{labelColumn}{TODO: description needed}

\item{id}{TODO: description needed}

\item{ageClassCutOffs}{TODO: description needed}

\item{ageClasses}{TODO: description needed}

\item{sppEquivCol}{TODO: description needed}

\item{sppEquiv}{TODO: description needed}

\item{crop2poly}{logical indicating whether to crop/mask `vtm` and `tsf` rasters to `poly`.}
}
\description{
TODO: needs description
}
