% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CASFRIv4.R
\name{prepSpeciesLayers_CASFRI}
\alias{prepSpeciesLayers_CASFRI}
\title{Prepare species layers from CASFRI v4}
\usage{
prepSpeciesLayers_CASFRI(
  destinationPath,
  outputPath,
  url = NULL,
  studyArea,
  rasterToMatch,
  sppEquiv,
  sppEquivCol,
  ...
)
}
\arguments{
\item{destinationPath}{path to data directory where objects will be downloaded or saved to}

\item{outputPath}{character, specifying the output directory to use}

\item{url}{if \code{NULL}, the default, use the default source url}

\item{studyArea}{A \verb{SpatialPolygons*} object used as the principle study region,
passed to \code{\link[reproducible:prepInputs]{reproducible::prepInputs()}}.}

\item{rasterToMatch}{A \code{RasterLayer} objects to use as the template for all subsequent
raster operations (i.e., the one used throughout the simulation).}

\item{sppEquiv}{table with species name equivalencies between the kNN and final naming formats.
See \code{data("sppEquivalencies_CA", "LandR")}.
For functions that have \code{mixedType}, this only necessary if \code{mixedType == 2}.
If not provided and \code{mixedType == 2}, will attempt to use \code{data("sppEquivalencies_CA", "LandR")}.}

\item{sppEquivCol}{the column name to use from \code{sppEquiv}.
For functions that have \code{mixedType}, only necessary if \code{mixedType == 2}.
If not provided and \code{mixedType == 2}, will attempt to use "Boreal".}

\item{...}{other arguments, used for compatibility with other \code{prepSpeciesLayers} functions.}
}
\description{
Prepare species layers from CASFRI v4
}
