% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterListByPoly.R
\name{rasterListByPoly}
\alias{rasterListByPoly}
\title{Create a list of rasters in each rep, at each timestep, for each polygon area}
\usage{
rasterListByPoly(files, polys, names, col, filter)
}
\arguments{
\item{files}{character vector giving paths to raster files}

\item{polys}{polygon object of class \code{sf}}

\item{names}{character vector giving the names of each of the subpolygons in \code{poly}}

\item{col}{character string giving the column name in \code{poly} to use}

\item{filter}{regex string giving partial filename in \code{files} to be filtered out
(i.e., when extracting reps, times, etc. from filenames)
(e.g., \code{'rstTimeSinceFire_'}, \code{'vegTypeMap_'})}
}
\value{
list of \code{RasterLayer} objects with attributes \code{reps}, \code{times}, \code{polyNames}
}
\description{
Create a list of rasters in each rep, at each timestep, for each polygon area
}
\note{
uses \code{future_lapply} internally to loop across \code{files};
set e.g., option \code{future.availableCores.fallback} appropriately for your system.
}
